:art: 동기2차
