+++ docker-compose-lecture.yml
@@ -7,19 +7,19 @@ volumes:
 services:
   member:
     build: member-domain
-    mem_limit: 400m
+    mem_limit: 300m
     ports:
-      - "8082:8080"
+      - "8082:80"
     environment:
-      - SPRING_PROFILES_ACTIVE=docker
+      - SPRING_PROFILES_3ACTIVE=docker
     depends_on:
       - kafka, mysql, gateway
   lecture:
     build: lecture-domain
-    mem_limit: 400m
+    mem_limit: 550M
     ports:
       - "8000:8000"
-      - "8081:8080"
+      - "8081:80"
     environment:
       - SPRING_PROFILES_ACTIVE=docker
     depends_on:
