+++ lecture-domain/Dockerfile
@@ -7,7 +7,7 @@ ENV TZ=Asia/Seoul
 
 RUN ln -snf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
 
-ENV JAVA_OPTS=""
+ENV JAVA_OPTS="-Xmx512m"
 
 ENTRYPOINT ["java","-agentlib:jdwp=transport=dt_socket,address=*:8000,server=y,suspend=n","-Djava.security.egd=file:/dev/./urandom","-jar","/app.jar"]
 #-Djava.security.egd=file:/dev/./urandom -Dspring.profiles.active=docker
