--- lecture-domain/Dockerfile copy
@@ -1,7 +1,0 @@
-FROM openjdk:12.0.2
-
-EXPOSE 8080
-
-ADD ./build/libs/*.jar app.jar
-
-ENTRYPOINT ["java","-jar","/app.jar"]
