+++ lecture-domain/src/main/java/lecturemgt/LectureApplication.java
@@ -3,7 +3,6 @@ package lecturemgt;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.context.ApplicationContext;
 
@@ -12,7 +11,7 @@ import lecturemgt.config.kafka.KafkaProcessor;
 
 @SpringBootApplication(scanBasePackages={"lecturemgt"})
 @EnableBinding(KafkaProcessor.class)
-@EnableFeignClients
+//@EnableFeignClients
 @EnableAutoConfiguration
 public class LectureApplication {
