+++ lecture-domain/src/main/java/lecturemgt/controller/dto/LecturesPostInDTO.java
@@ -1,4 +1,4 @@
-package lecturemgt.web.rest.dto;
+package lecturemgt.controller.dto;
 
 import lombok.Getter;
 import lombok.Setter;
