+++ lecture-domain/src/main/java/lecturemgt/controller/errors/FeignClientException.java
@@ -1,4 +1,4 @@
-package lecturemgt.web.rest.errors;
+package lecturemgt.controller.errors;
