+++ lecture-domain/src/main/java/lecturemgt/controller/errors/FeignClientExceptionErrorDecoder.java
@@ -1,4 +1,4 @@
-package lecturemgt.web.rest.errors;
+package lecturemgt.controller.errors;
 
 
 import feign.Response;
