+++ lecture-domain/src/main/java/lecturemgt/controller/mapper/LecturePOSTOutMapper.java
@@ -1,21 +1,20 @@
-package lecturemgt.web.rest.mapper;
+package lecturemgt.controller.mapper;
 
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
+import lecturemgt.controller.dto.LecturesPostOutDTO;
 import lecturemgt.domain.Lecture;
-import lecturemgt.web.rest.dto.LecturesPostInDTO;
 
 
 @Mapper(componentModel = "spring", uses = {})
-public interface LecturePOSTInMapper extends EntityMapper<LecturesPostInDTO, Lecture> {
+public interface LecturePOSTOutMapper extends EntityMapper<LecturesPostOutDTO, Lecture> {
 
-    @Mapping(target = "id", ignore = true)
-    @Mapping(target = "categoryName", ignore = true)
+
     @Mapping(target = "status", ignore = true)
     @Mapping(target = "version", ignore = true)
-    Lecture toEntity(LecturesPostInDTO lectureDTO);
+    Lecture toEntity(LecturesPostOutDTO lectureDTO);
 
     default Lecture fromId(Long id) {
         if (id == null) {
