+++ lecture-domain/src/main/java/lecturemgt/domain/Lecture.java
@@ -1,8 +1,5 @@
 package lecturemgt.domain;
 
-import javax.persistence.AttributeOverride;
-import javax.persistence.Column;
-import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
