--- lecture-domain/src/main/java/lecturemgt/infra/LectureController.java
@@ -1,19 +1,0 @@
-//package lecturemgt.infra;
-//
-//import javax.transaction.Transactional;
-//
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.web.bind.annotation.RequestMapping;
-//import org.springframework.web.bind.annotation.RestController;
-//
-//import lecturemgt.domain.LectureRepository;
-//
-//@RestController
-//@RequestMapping(value = "/lectures")
-//@Transactional
-//public class LectureController {
-//
-//    @Autowired
-//    LectureRepository lectureRepository;
-//    // keep
-//}
