+++ lecture-domain/src/main/resources/application.yml
@@ -43,7 +43,7 @@ logging:
 
 spring.profiles: docker
 
-server.port: 8080
+server.port: 80
 
 spring:
   cloud:
