+++ member-domain/Dockerfile
@@ -3,4 +3,4 @@ COPY target/*SNAPSHOT.jar app.jar
 EXPOSE 80
 ENV TZ=Asia/Seoul
 RUN ln -snf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
-ENTRYPOINT ["java","-Xmx400M","-Djava.security.egd=file:/dev/./urandom","-jar","/app.jar","--spring.profiles.active=docker"]
+ENTRYPOINT ["java","-Xmx200M","-Djava.security.egd=file:/dev/./urandom","-jar","/app.jar","--spring.profiles.active=docker"]
